fix issue: 비로그인 상태 시 전체보기 화면 에러 해결, 그룹 프로필 상세보기 막기, api matchPage 대신 index에서 호출
